Fix faulty props
The `withDefaultProps` higher order component is trying to do something clever with the types, but what happened that defaultProps had `{}` as a type for map options, instead of Wrld.MapOptions, causing the exported component to take a generic object as the option, rather thant the proper type.

Correctly typing defaultProps as a partial version of WrldMapProps guarantees that all its field have the correct types, also highlighting an awkward typo.